backend/requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
whitenoise>=6.6.0

# Database
psycopg2-binary>=2.9.9
djongo>=1.3.6
mongoengine>=0.27.0
pymongo>=4.6.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# Environment
python-dotenv>=1.0.0

# Payments
stripe>=7.0.0

# Utils
celery>=5.3.0
redis>=5.0.0
pillow>=10.0.0

# Development
black>=23.0.0
flake8>=6.1.0
pytest>=7.4.0
pytest-django>=4.5.0

# Production
gunicorn>=21.2.0
